pyproject.toml
[tool.poetry]
name = "deposit"
version = "0.1.0"
description = ""
authors = ["Aleksandr <ShirshovAleksandrWA@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
python-dateutil = "^2.9.0.post0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[tool.poetry.dev-dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
pytest-django = "^4.8.0"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
